Separate VISCA implementation into multiple files
Since the source code `ptz-visca.cpp` is increasing, I'd like to suggest
to move different implementations into different files.
