+++ ptz-visca.cpp
@@ -11,9 +11,6 @@
 #include "ptz-visca.hpp"
 #include <util/base.h>
 
-std::map<QString, ViscaUART*> ViscaUART::interfaces;
-std::map<int, ViscaUDPSocket*> ViscaUDPSocket::interfaces;
-
 /* Visca specific datagram field classes */
 class visca_u4 : public int_field {
 public:
@@ -440,12 +437,6 @@ const PTZCmd VISCA_PanTilt_LimitClearDownLeft("810106070100070f0f0f070f0f0fff", 
 		new visca_u16("tilt_limit_down", 10)
 	});
 
-#define VISCA_RESPONSE_ADDRESS   0x30
-#define VISCA_RESPONSE_ACK	 0x40
-#define VISCA_RESPONSE_COMPLETED 0x50
-#define VISCA_RESPONSE_ERROR     0x60
-#define VISCA_PACKET_SENDER(pkt) ((unsigned)((pkt)[0] & 0x70) >> 4)
-
 const QMap<int, std::string> PTZVisca::viscaVendors = {
 	{ 0x0001, "Sony" },
 	{ 0x0109, "Birddog" },
@@ -703,422 +694,4 @@ void PTZVisca::memory_set(int i)
 void PTZVisca::memory_recall(int i)
 {
 	send(VISCA_CAM_Memory_Recall, {i});
-}
-
-/*
- * VISCA over serial UART implementation
- */
-ViscaUART::ViscaUART(QString &port_name) :
-	PTZUARTWrapper(port_name)
-{
-	camera_count = 0;
-}
-
-bool ViscaUART::open()
-{
-	camera_count = 0;
-	bool rc = PTZUARTWrapper::open();
-	if (rc)
-		send(VISCA_ENUMERATE.cmd);
-	return rc;
-}
-
-void ViscaUART::close()
-{
-	PTZUARTWrapper::close();
-	camera_count = 0;
-}
-
-void ViscaUART::receive_datagram(const QByteArray &packet)
-{
-	ptz_debug("VISCA <-- %s", packet.toHex(':').data());
-	if (packet.size() < 3)
-		return;
-	if ((packet[1] & 0xf0) == VISCA_RESPONSE_ADDRESS) {
-		switch (packet[1] & 0x0f) { /* Decode Packet Socket Field */
-		case 0:
-			camera_count = (packet[2] & 0x7) - 1;
-			blog(LOG_INFO, "VISCA Interface %s: %i camera%s found",
-				qPrintable(uart.portName()),
-				camera_count, camera_count == 1 ? "" : "s");
-			send(VISCA_IF_CLEAR.cmd);
-			emit reset();
-			break;
-		case 1:
-			// Response from IF_CLEAR message; ignore
-			break;
-		case 8:
-			/* network change, trigger a change */
-			send(VISCA_ENUMERATE.cmd);
-			break;
-		default:
-			break;
-		}
-		return;
-	}
-
-	emit receive(packet);
-}
-
-void ViscaUART::receiveBytes(const QByteArray &msg)
-{
-	for (auto b : msg) {
-		rxbuffer += b;
-		if ((b & 0xff) == 0xff) {
-			if (rxbuffer.size())
-				receive_datagram(rxbuffer);
-			rxbuffer.clear();
-		}
-	}
-}
-
-ViscaUART * ViscaUART::get_interface(QString port_name)
-{
-	ViscaUART *iface;
-	ptz_debug("Looking for UART object %s", qPrintable(port_name));
-	iface = interfaces[port_name];
-	if (!iface) {
-		ptz_debug("Creating new VISCA object %s", qPrintable(port_name));
-		iface = new ViscaUART(port_name);
-		iface->open();
-		interfaces[port_name] = iface;
-	}
-	return iface;
-}
-
-PTZViscaSerial::PTZViscaSerial(OBSData config)
-	: PTZVisca(config), iface(NULL)
-{
-	set_config(config);
-	auto_settings_filter += {"port", "address", "baud_rate"};
-}
-
-PTZViscaSerial::~PTZViscaSerial()
-{
-	attach_interface(nullptr);
-}
-
-void PTZViscaSerial::attach_interface(ViscaUART *new_iface)
-{
-	if (iface)
-		iface->disconnect(this);
-	iface = new_iface;
-	if (iface) {
-		connect(iface, &ViscaUART::receive, this, &PTZViscaSerial::receive);
-		connect(iface, &ViscaUART::reset, this, &PTZViscaSerial::reset);
-	}
-}
-
-void PTZViscaSerial::reset()
-{
-	cmd_get_camera_info();
-}
-
-void PTZViscaSerial::send_immediate(const QByteArray &msg_)
-{
-	QByteArray msg = msg_;
-	msg[0] = (char)(0x80 | address & 0x7); // Set the camera address
-	iface->send(msg);
-}
-
-void PTZViscaSerial::set_config(OBSData config)
-{
-	PTZDevice::set_config(config);
-	const char *uart = obs_data_get_string(config, "port");
-	address = qBound(1, (int)obs_data_get_int(config, "address"), 7);
-	if (!uart)
-		return;
-
-	iface = ViscaUART::get_interface(uart);
-	iface->setConfig(config);
-	attach_interface(iface);
-}
-
-OBSData PTZViscaSerial::get_config()
-{
-	OBSData config = PTZDevice::get_config();
-	obs_data_apply(config, iface->getConfig());
-	obs_data_set_int(config, "address", address);
-	return config;
-}
-
-obs_properties_t *PTZViscaSerial::get_obs_properties()
-{
-	obs_properties_t *props = PTZVisca::get_obs_properties();
-	obs_property_t *p = obs_properties_get(props, "interface");
-	obs_properties_t *config = obs_property_group_content(p);
-	obs_property_set_description(p, "VISCA Connection");
-
-	iface->addOBSProperties(config);
-	obs_properties_add_int(config, "address", "VISCA ID", 1, 7, 1);
-
-	return props;
-}
-
-/*
- * VISCA over IP implementation
- */
-ViscaUDPSocket::ViscaUDPSocket(int port) :
-	visca_port(port)
-{
-	if (!visca_socket.bind(QHostAddress::Any, visca_port)) {
-		blog(LOG_INFO, "VISCA-over-IP bind to port %i failed", visca_port);
-		return;
-	}
-	connect(&visca_socket, &QUdpSocket::readyRead, this, &ViscaUDPSocket::poll);
-}
-
-void ViscaUDPSocket::receive_datagram(QNetworkDatagram &dg)
-{
-	QByteArray data = dg.data();
-	int type = ((data[0] & 0xff) << 8) | (data[1] & 0xff);
-	int size = data[2] << 8 | data[3];
-	int seq = data[4] << 24 | data[5] << 16 | data[6] << 8 | data[7];
-
-	if ((data.size() != size + 8) || size < 1) {
-		ptz_debug("VISCA UDP (malformed) <-- %s", qPrintable(data.toHex(':')));
-		return;
-	}
-	ptz_debug("VISCA UDP type=%.4x seq=%i size=%i <-- %s",
-		type, seq, size, qPrintable(data.toHex(':')));
-
-	switch (type) {
-	case 0x0111:
-		emit receive(data.mid(8, size));
-		break;
-	case 0x0200:
-	case 0x0201: /* Check for sequence number out of sync */
-		if (data[8] == (char)0x0f && data[8+1] == (char)1)
-			emit reset();
-		break;
-	default:
-		ptz_debug("VISCA UDP unrecognized type: %x", type);
-	}
-
-}
-
-void ViscaUDPSocket::send(QHostAddress ip_address, const QByteArray &packet)
-{
-	ptz_debug("VISCA UDP --> %s", qPrintable(packet.toHex(':')));
-	visca_socket.writeDatagram(packet, ip_address, visca_port);
-}
-
-void ViscaUDPSocket::poll()
-{
-	while (visca_socket.hasPendingDatagrams()) {
-		QNetworkDatagram dg = visca_socket.receiveDatagram();
-		receive_datagram(dg);
-	}
-}
-
-ViscaUDPSocket * ViscaUDPSocket::get_interface(int port)
-{
-	ViscaUDPSocket *iface;
-	ptz_debug("Looking for Visca UDP Socket object %i", port);
-	iface = interfaces[port];
-	if (!iface) {
-		ptz_debug("Creating new VISCA object %i", port);
-		iface = new ViscaUDPSocket(port);
-		interfaces[port] = iface;
-	}
-	return iface;
-}
-
-PTZViscaOverIP::PTZViscaOverIP(OBSData config)
-	: PTZVisca(config), iface(NULL)
-{
-	address = 1;
-	set_config(config);
-	auto_settings_filter += {"port", "address"};
-}
-
-PTZViscaOverIP::~PTZViscaOverIP()
-{
-	attach_interface(nullptr);
-}
-
-void PTZViscaOverIP::attach_interface(ViscaUDPSocket *new_iface)
-{
-	if (iface)
-		iface->disconnect(this);
-	iface = new_iface;
-	if (iface) {
-		connect(iface, &ViscaUDPSocket::receive, this, &PTZViscaOverIP::receive);
-		connect(iface, &ViscaUDPSocket::reset, this, &PTZViscaOverIP::reset);
-		reset();
-	}
-}
-
-void PTZViscaOverIP::reset()
-{
-	sequence = 1;
-	iface->send(ip_address, QByteArray::fromHex("020000010000000001"));
-	send(VISCA_Clear);
-	cmd_get_camera_info();
-}
-
-void PTZViscaOverIP::send_immediate(const QByteArray &msg)
-{
-	QByteArray p = QByteArray::fromHex("0100000000000000") + msg;
-	p[1] = (0x9 == msg[1]) ? 0x10 : 0x00;
-	p[3] = msg.size();
-	p[4] = (sequence >> 24) & 0xff;
-	p[5] = (sequence >> 16) & 0xff;
-	p[6] = (sequence >> 8) & 0xff;
-	p[7] = sequence & 0xff;
-	p[8] = '\x81';
-	sequence++;
-	iface->send(ip_address, p);
-}
-
-void PTZViscaOverIP::set_config(OBSData config)
-{
-	PTZDevice::set_config(config);
-	const char *ip = obs_data_get_string(config, "address");
-	if (ip)
-		ip_address = QHostAddress(ip);
-	int port = obs_data_get_int(config, "port");
-	if (!port)
-		port = 52381;
-	attach_interface(ViscaUDPSocket::get_interface(port));
-}
-
-OBSData PTZViscaOverIP::get_config()
-{
-	OBSData config = PTZDevice::get_config();
-	obs_data_set_string(config, "address", qPrintable(ip_address.toString()));
-	obs_data_set_int(config, "port", iface->port());
-	return config;
-}
-
-obs_properties_t *PTZViscaOverIP::get_obs_properties()
-{
-	obs_properties_t *props = PTZVisca::get_obs_properties();
-	obs_property_t *p = obs_properties_get(props, "interface");
-	obs_properties_t *config = obs_property_group_content(p);
-	obs_property_set_description(p, "VISCA-over-IP Connection");
-	obs_properties_add_text(config, "address", "IP Address", OBS_TEXT_DEFAULT);
-	obs_properties_add_int(config, "port", "UDP port", 1, 65535, 1);
-	return props;
-}
-
-/*
- * VISCA over TCP implementation (e.g., PTZOptics cameras
- */
-PTZViscaOverTCP::PTZViscaOverTCP(OBSData config)
-	: PTZVisca(config)
-{
-	address = 1;
-	set_config(config);
-	auto_settings_filter += {"port", "host"};
-	visca_socket.setSocketOption(QAbstractSocket::KeepAliveOption, 1);
-	connect(&visca_socket, &QTcpSocket::readyRead, this, &PTZViscaOverTCP::poll);
-	connect(&visca_socket, &QTcpSocket::stateChanged, this, &PTZViscaOverTCP::on_socket_stateChanged);
-}
-
-void PTZViscaOverTCP::reset()
-{
-	send(VISCA_Clear);
-	cmd_get_camera_info();
-}
-
-void PTZViscaOverTCP::connectSocket()
-{
-	visca_socket.connectToHost(host, port);
-}
-
-void PTZViscaOverTCP::on_socket_stateChanged(QAbstractSocket::SocketState state)
-{
-	blog(LOG_INFO, "VISCA_over_TCP socket state: %s", qPrintable(QVariant::fromValue(state).toString()));
-	switch (state) {
-	case QAbstractSocket::UnconnectedState:
-		/* Attempt reconnection periodically */
-		QTimer::singleShot(900, this, SLOT(connectSocket));
-		break;
-	case QAbstractSocket::ConnectedState:
-		blog(LOG_INFO, "VISCA_over_TCP %s connected", QT_TO_UTF8(objectName()));
-		reset();
-		break;
-	default:
-		break;
-	}
-}
-
-void PTZViscaOverTCP::send_immediate(const QByteArray &msg)
-{
-	if (visca_socket.state() == QAbstractSocket::UnconnectedState)
-		connectSocket();
-	ptz_debug("VISCA_over_TCP --> %s", qPrintable(msg.toHex(':')));
-	visca_socket.write(msg);
-}
-
-void PTZViscaOverTCP::receive_datagram(const QByteArray &packet)
-{
-	int camera_count = 0;
-	ptz_debug("VISCA_over_TCP <-- %s", packet.toHex(':').data());
-	if (packet.size() < 3)
-		return;
-	if ((packet[1] & 0xf0) == VISCA_RESPONSE_ADDRESS) {
-		switch (packet[1] & 0x0f) { /* Decode Packet Socket Field */
-		case 0:
-			camera_count = (packet[2] & 0x7) - 1;
-			blog(LOG_INFO, "VISCA-over-TCP Interface %i camera%s found",
-				camera_count, camera_count == 1 ? "" : "s");
-			send_immediate(VISCA_IF_CLEAR.cmd);
-			reset();
-			break;
-		case 1:
-			// Response from IF_CLEAR message; ignore
-			break;
-		case 8:
-			/* network change, trigger a change */
-			send_immediate(VISCA_ENUMERATE.cmd);
-			break;
-		default:
-			break;
-		}
-		return;
-	}
-	receive(packet);
-}
-
-void PTZViscaOverTCP::poll()
-{
-	for (auto b : visca_socket.readAll()) {
-		rxbuffer += b;
-		if ((b & 0xff) == 0xff) {
-			if (rxbuffer.size())
-				receive_datagram(rxbuffer);
-			rxbuffer.clear();
-		}
-	}
-}
-
-void PTZViscaOverTCP::set_config(OBSData config)
-{
-	PTZDevice::set_config(config);
-	host = obs_data_get_string(config, "host");
-	port = obs_data_get_int(config, "port");
-	if (!port)
-		port = 5678;
-	connectSocket();
-}
-
-OBSData PTZViscaOverTCP::get_config()
-{
-	OBSData config = PTZDevice::get_config();
-	obs_data_set_string(config, "host", QT_TO_UTF8(host));
-	obs_data_set_int(config, "port", port);
-	return config;
-}
-
-obs_properties_t *PTZViscaOverTCP::get_obs_properties()
-{
-	obs_properties_t *props = PTZVisca::get_obs_properties();
-	obs_property_t *p = obs_properties_get(props, "interface");
-	obs_properties_t *config = obs_property_group_content(p);
-	obs_property_set_description(p, "VISCA-over-IP Connection");
-	obs_properties_add_text(config, "host", "IP Host", OBS_TEXT_DEFAULT);
-	obs_properties_add_int(config, "port", "TCP port", 1, 65535, 1);
-	return props;
 }
